refactor(gui): clean up
