+++ src/guinea/components/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod general;
+pub(crate) mod machine;
